Revert "done building singup form"
