+++ package.json
@@ -10,20 +10,17 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@hookform/resolvers": "^5.0.1",
     "@tailwindcss/vite": "^4.0.15",
     "framer-motion": "^12.6.5",
     "lucide-react": "^0.483.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
-    "react-hook-form": "^7.56.1",
     "react-lottie-player": "^2.1.0",
     "react-tsparticles": "^2.12.2",
     "styled-components": "^6.1.17",
     "tailwindcss": "^4.0.15",
     "tsparticles-engine": "^2.12.0",
-    "tsparticles-slim": "^2.12.0",
-    "zod": "^3.24.3"
+    "tsparticles-slim": "^2.12.0"
   },
   "devDependencies": {
     "@eslint/js": "^9.21.0",
@@ -35,7 +32,6 @@
     "eslint-plugin-react-hooks": "^5.1.0",
     "eslint-plugin-react-refresh": "^0.4.19",
     "globals": "^15.15.0",
-    "react-router-dom": "^7.5.3",
     "typescript": "~5.7.2",
     "typescript-eslint": "^8.24.1",
     "vite": "^6.2.0"
