--- src/components/atoms/Input.tsx
@@ -1,52 +1,0 @@
-import React, { forwardRef } from 'react';
-
-// Icons object with SVG paths
-const icons = {
-  user: (
-    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" className="w-4 h-4">
-      <path d="M20 21v-2a4 4 0 0 0-4-4H8a4 4 0 0 0-4 4v2"></path>
-      <circle cx="12" cy="7" r="4"></circle>
-    </svg>
-  ),
-  mail: (
-    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" className="w-4 h-4">
-      <path d="M4 4h16c1.1 0 2 .9 2 2v12c0 1.1-.9 2-2 2H4c-1.1 0-2-.9-2-2V6c0-1.1.9-2 2-2z"></path>
-      <polyline points="22,6 12,13 2,6"></polyline>
-    </svg>
-  ),
-  lock: (
-    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" className="w-4 h-4">
-      <rect x="3" y="11" width="18" height="11" rx="2" ry="2"></rect>
-      <path d="M7 11V7a5 5 0 0 1 10 0v4"></path>
-    </svg>
-  ),
-};
-
-type IconType = keyof typeof icons;
-
-interface InputProps extends React.InputHTMLAttributes<HTMLInputElement> {
-  icon?: IconType;
-}
-
-const Input = forwardRef<HTMLInputElement, InputProps>(
-  ({ icon, className = '', ...props }, ref) => {
-    return (
-      <div className="relative w-full">
-        {icon && (
-          <div className="absolute left-3 top-1/2 -translate-y-1/2 text-gray-400">
-            {icons[icon]}
-          </div>
-        )}
-        <input
-          ref={ref}
-          className={`w-full px-4 py-3 ${icon ? 'pl-10' : ''} bg-gray-800/80 border border-gray-700 rounded-md text-white text-sm placeholder-gray-500 focus:outline-none focus:ring-1 focus:ring-cyan-400 focus:border-cyan-400 transition duration-200 ${className}`}
-          {...props}
-        />
-      </div>
-    );
-  }
-);
-
-Input.displayName = 'Input';
-
-export default Input;
